NODE-517: Fix test block parsing to work with nested structures and look for underscored field name.
